requirements.txt
Flask==2.0.1
gunicorn==20.1.0
reportlab==3.5.68
python-docx==0.8.11
